Add support for iOS/macCatalyst and improve protocol
The STAnnotationsDataSource and STLineAnnotation protocols are updated to
better support iOS/macCatalyst in addition to macOS. The typealias for View is
removed in favor of explicitly using NSView or UIView depending on the
platform. The STLineAnnotation protocol is also updated to use the new 'any'
keyword for the location property type.